requirements.txt
beautifulsoup4==4.4.1
Django==1.9.7
django-appconf==1.0.2
django-imagekit==3.3
django-modelcluster==2.0
django-taggit==0.18.3
django-treebeard==4.0.1
djangorestframework==3.3.3
html5lib==0.9999999
pilkit==1.1.13
Pillow==3.3.0
pytz==2016.4
six==1.10.0
Unidecode==0.4.19
wagtail==1.5.2
Willow==0.3.1
